fix: (processm.services) changing data store size from int to long (#309)

Closes #309

See merge request processm-team/processm!317

Former-commit-id: 566fdb52e86d0032e35760c6ce5e727c7bccea9e

+++ processm.services/src/main/kotlin/processm/services/api/DataStoresApi.kt
@@ -74,7 +74,7 @@ fun Route.DataStoresApi() {
                 DataStore(
                     dataStore.name,
                     dataStore.id.value,
-                    dataStoreSize.toInt(),
+                    dataStoreSize,
                     dataStore.creationDate
                 )
             )
